save/download in snapshot util
