+++ client/utils/draw.js
@@ -31,8 +31,6 @@ import KalmanFilter from 'kalmanjs'
 const xFilter = new KalmanFilter()
 const yFilter = new KalmanFilter()
 
-let fullImageStr
-
 export function createProject(window, cnv) {
   paper.install(window)
   paper.setup(cnv)
@@ -43,27 +41,6 @@ export async function clearCanvas() {
 
   //restarts drawing line, also needed as bugfix
   store.dispatch(drawOff())
-}
-
-export function saveCanvas() {
-  const backgroundCanvas = document.getElementById('background')
-  const bgCtx = backgroundCanvas.getContext('2d')
-  const canvas = document.getElementById('output')
-
-  bgCtx.drawImage(canvas, 0, 0)
-  fullImageStr = bgCtx.canvas.toDataURL('image/png')
-
-  return fullImageStr
-}
-
-//create a DOM element to hold download ref
-export function download() {
-  //create a DOM element to hold download ref
-  let element = document.createElement('a')
-  const file = fullImageStr.replace('image/png', 'image/octet-stream')
-  element.href = file
-  element.download = 'airbrush.png'
-  element.click()
 }
 
 function getColor() {
